+++ react/components/WordpressProductSearchResult.tsx
@@ -7,7 +7,7 @@ import React, {
   useEffect,
   useRef,
 } from 'react'
-import { defineMessages } from 'react-intl'
+import { defineMessages, useIntl } from 'react-intl'
 import { useQuery } from 'react-apollo'
 import { Spinner, Pagination } from 'vtex.styleguide'
 import { useCssHandles } from 'vtex.css-handles'
@@ -15,6 +15,7 @@ import { useCssHandles } from 'vtex.css-handles'
 import WordpressTeaser from './WordpressTeaser'
 import withSearchContext from './withSearchContext'
 import SearchPosts from '../graphql/SearchPosts.graphql'
+import Settings from '../graphql/Settings.graphql'
 
 interface Props {
   searchQuery: any
@@ -43,9 +44,12 @@ const WordpressSearchResult: StorefrontFunctionComponent<Props> = ({
   subcategoryUrls,
   postsPerPage,
 }) => {
+  const intl = useIntl()
   const [page, setPage] = useState(1)
   const [perPage, setPerPage] = useState(postsPerPage)
+  const [selectedOption, setSelectedOption] = useState(postsPerPage)
   const handles = useCssHandles(CSS_HANDLES)
+  const { data: dataS } = useQuery(Settings)
   const { loading, error, data, fetchMore } = useQuery(SearchPosts, {
     skip: !searchQuery,
     variables: {
@@ -84,12 +88,19 @@ const WordpressSearchResult: StorefrontFunctionComponent<Props> = ({
         postsPerPage * 3,
         postsPerPage * 4,
       ]}
+      selectedOption={selectedOption}
       currentItemFrom={(page - 1) * perPage + 1}
       currentItemTo={page * perPage}
       textOf="of"
-      textShowRows="posts per page"
+      textShowRows={
+        dataS?.appSettings?.displayShowRowsText === false
+          ? null
+          : // eslint-disable-next-line @typescript-eslint/no-use-before-define
+            intl.formatMessage(messages.postsPerPage)
+      }
       totalItems={data?.wpPosts?.total_count ?? 0}
       onRowsChange={({ target: { value } }: ChangeEvent<HTMLInputElement>) => {
+        setSelectedOption(+value)
         setPage(1)
         setPerPage(+value)
         fetchMore({
@@ -217,6 +228,10 @@ const WordpressSearchResult: StorefrontFunctionComponent<Props> = ({
 }
 
 const messages = defineMessages({
+  postsPerPage: {
+    defaultMessage: 'posts per page',
+    id: 'store/wordpress-integration.wordpressPagination.postsPerPage',
+  },
   title: {
     defaultMessage: '',
     id: 'admin/editor.wordpressSearchResult.title',